Discard async proxies when asyncLoadAll() throws exception
When AsyncLoadingCache#getAll() bulk loads, it first inserts proxies
into the cache for each missing entry. It then calls the loader's
asyncLoadAll() and populates the proxies with the results. If that
method instead throws an exception or returns null, the proxies
should be removed before propogating the error. Previously these
remained in the cache, unfulfilled and thus blocking consumers
forever.