Prototype of hierarchical timer wheel
A timer wheel allows timers to be added, removed, or fired in amortized
O(1) time. Instead of using a O(lg n) priority queue, timers are
assigned to a bucket by hashing. Each wheel represents a resolution,
e.g. seconds / minutes / hours, and structured into a hierarchy. As the
wheels rotate, events cascade from a higher level to the lower. When a
bucket then all of its timers are evaluated, either cascaded or fired.
This amortizes the cost as cascading is done in small batches.

This serves as the foundation for variable expiration, where some
entries are short lived and others long lived. The custom policy is
defined by the user and the timer wheel is operated on during the
cache maintenance cycles.

This is work in progress. But shows the general idea of where we are
going.