Add Gradle build cache to the simulator
This is a 30 day trace of Gradle's distributed build cache. This may
only be what the server's see and that the client caches absorb most
of the requests. This would explain why LRU, FIFO, and Random are near
optimal while LFU is very poor.

Thanks to the Gradle team for extracting this from their logs.