Use large sample and pivot for hill climbers
A small sample is reactive and amortizes the movement by shifting only
a few entries between lists. But it also suffers from noise. A large
sample does not have noise, but shifts many entries at once when
reconfiguring. This changes the config to favor that approach since it
provides a better hit rate. We will probably make the movement penalty
amortized by spreading it out during the sample period.