Improved default layering rules for JCache configuration (fixes #171)
Previously defaults were only honored for resource configured caches. An
explicit cache configuration (via CacheManager#createCache) was not did
not use the defaults for non-JSR settings. This made default a poorly
named template only used by resource configurations.

It is now used in these cases as well, which reduces surprises. A nice
benefit is that anonymous caches created by the annotations are not
stuck on the JSR's defaults (unbounded, serialized). A safe threshold
can be set in the defaults to avoid memory leaks, etc.

Restricted CacheManager#createCache from overriding the externalized
configuration. As the caches are lazily loaded or destroyed, creating
a defined cache could be forced to a setting. That caused confusion on
when the layering, e.g. would the explicit layer on the externalized?

Added missing key-type and value-type configuration settings for getting
a cache by its name+types.