Support putting a mapping with variable expiration (fixes #163)
The policy specific extension, Policy.VarExpiration, can now ignore
the configured Expiry and add a mapping with the supplied duration.
This better supports custom cache APIs that expose this, often due
to abstracting memcached. An Expiry must be configured to handle a
computation, but the new put(k, v, duration, unit) will ignore it
when the write is successful.