Remove the simulator's split package for Jigsaw modularity
The CountMin64 implementation is provided by AddThis StreamLib, but
required modifying the package-private state for the TinyLFU reset
operation. This was done by using the same package name for field
access, but will break in JDK9 modularity restrictions.

Now the CountMin64 is a slimmed down fork of the original code. The
BloomFilter option was removed (since 2 others are provided) so that
the dependency could be removed. The JavaDoc retains the attribution.