Use double math for gradient descent steps to avoid rounding errors
Removed the "mayAdapt" and hit density logic. The problems were due to
and old hack to deal with warming up an empty cache. In a loopy trace its
okay to have full misses, and this confused SGD. There's still some quirks
to resolve, but its slowly getting closer to correct.

Cleaned up climbers to be friendlier for future refactorings.