Minor test gc optimization
Tests are hitting the time limit due to the exhaustive nature of brute
forcing with millions of cases. This change avoids some of the
unnecessary churn, though a deeper dive is needed or a request to extend
the free limit.

Using ParallelGC is only required for deterministic soft reference
collection in tests. Switching the other tests to use G1 reduces their
tasks by 1.5 minutes each.