Add fast-filter, snia msr trace, improved type checking
This contains a lot of small changes that accumulated locally, but none
are complex or very interesting. Therefore work to split into multiple
commits wasn't done.

Fixes #320 by adding ErrorProne's `@CompatibleWith` annotation on
`getIfPresent` and `invalidate`. These take object keys, for subtle
reasons, and therefore lets developers mistakenly supply the wrong
parameter. If using ErrorProne a warning will be shown.

SNIA is a popular repository of trace files, such as the MSR Cambridge
trace. However the most cited traces are historical and SNIA strongly
warns against using. This is included for comparing against papers
which ignored this concern and that traces are hard to come by.

Adopted the Gradle nullaway plugin, updated dependencies, and fixed
warnings from the updated static analyzers. Updated to JCache 1.1.1
which changed what `CacheEntryEvent#getOldValue()` returns in the TCK.

Fixes #317 by adding Thomas Mueller's [FastFilter][filter] library to
the bloom filters available in the simulator. This provides a good
option to evaluate if the TinyLfu doorkeeper is adopted.

[filter]: https://github.com/FastFilter/fastfilter_java

Co-authored-by: Thomas Mueller <thomas.tom.mueller@gmail.com>