Make add/update race more benign
This is pretty hard to hit and isn't harmful, but caught when scanning
code and verified manually. Unfortunately there is no way to write an
explicit test case, only using the debugger works.

Imagine a weighted cache where the weight is the value.

T0. Add (A, 1) to hash table
T1. Update to (A, 20) in hash table, enqueue UpdateTask(A, (20-1))
T2. Adder (T0) enqueues AddTask(A, 1) to write buffer
T3. Both add/update operations complete, maintenance is scheduled
T4. Update task runs, sets PolicyWeight to 19
T5. Add task runs, sets PolicyWeight to 1

However, note that Node PolicyWeight and Weight are different
fields, and the weights are usually handled as deltas. Thus even
though the Node's policyWeight is wrong, the impact is very small
due to how other logic handles things.

To hit this problem you need two writers for the same key, with
different values, on a weighted cache to be thrashing it. If the
adder is context switched out before adding to the buffer, then
the inconsistency leaks.