JavaDoc that asMap() provides atomic computations
A map computation is a powerful and useful feature that has always been
fully supported. While not all ConcurrentMap implementations offer it
atomically, ConcurrentHashMap does and Caffeine extends that support.

Because compute/merge methods were added in Java 8 and are not always
atomic, users tend to not consider them. Since their characteristics is
not well-known or defined, it should be noted specifically in asMap().
As with any feature, it is useful but should not be abused (e.g. no
support for recursive computations).