Fix test case for scheduling the next timerWheel bucket
This shows why one does not debug tests when tired, on the train home. It
broke with the test (as is now) because minimum span when non-empty is
the bucket's. The node's lifetime isn't correct, as it may vary
significantly compared to neighbors due to the cascading effect.The global
minimum should be tested against to validate that the correct schedule is
determined.