Isolate osgi failure, due to guava's split packages
OSGi loads bundles as atomic units on a classpath, so split packages
are not supported. When the class verifier is enabled, it fails to
find required class dependencies and the test fails. When disabled
the test passes since we do not need to load the missing class.

Disabling verifying is a deprecated feature in JDK13 for removal. It
was originally added to speed up tests, since resolved by reflectively
loading classes in the factories rather than explicit class references.

https://github.com/google/guava/issues/3320