Improve countmin4 spread if hashCode is zero
When the hashCode is zero, then due to multiplicative spreads the array
indexes were also zero. Similarly the difference between positive and
negative numbers, which flip only the signum bit, had a higher chance
of collisions. Even with these caveats, the spreads were quite good and
the cache has other schemes to protect itself from HashDoS style of
attacks.

The improvement here is simply too easy to ignore for a nicer theoretical
advantage. By adding the seed, which differs per row, we avoid this eye
sore. The change in hit rates is noise (0.01-0.5 %, often better).